pyproject.toml
[tool.poetry]
name = "chat-test-task"
version = "0.1.0"
description = ""
authors = ["Egor Malov <egor.malov2014@gmail.com>"]
readme = "README.md"
packages = [{include = "chat_test_task"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
pre-commit = "^3.3.3"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
channels = "^4.0.0"
redis = "^4.5.5"
django-webpack-loader = "^2.0.1"
daphne = "^4.0.0"
channels-redis = "^4.1.0"
drf-yasg = "^1.21.6"
django-debug-toolbar = "^4.1.0"
pytils = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
